64bit: widen sub_10028650 bs-alias locals (a1, v1) to intptr_t
The IDA decomp stashed the bs pointer in two register-alias locals
typed as plain int.  My prior fix routed the AINode-Stand comparison
through BotAINode((bot_state_t *)v1) — but v1 was still int, so the
recast truncated bs to 32 bits and the macro indexed botainodes[]
with a wildly wrong (bs - botstates) result.  Crashed at
sub_10028650+0x2a4 on the ldr loading botainodes[index].

Callees taking bot_state_t * receive the real pointer; int-typed
callees still see the low 32 bits as before (unchanged behaviour for
the pre-existing 32-bit path).

Smoke: 120s with 3 bots, 30 chat lines, no crash.

Source-commit: 8d711015ee89a8865b1ef2cd71aa2eff9e4c1da9

+++ botlib/botlib.c
@@ -22037,8 +22037,11 @@ LABEL_64:
 //----- (10028650) --------------------------------------------------------
 void __cdecl sub_10028650(bot_state_t *bs)
 {
-  int a1 = (int)(intptr_t)bs;
-  int v1; // edi
+  /* a1/v1 hold the bs pointer; on 64-bit we must keep their full
+   * width or BotAINode((bot_state_t *)v1) re-truncates to a bogus
+   * index.  Keep both as intptr_t to preserve the original cast sites. */
+  intptr_t a1 = (intptr_t)bs;
+  intptr_t v1; // edi (was int)
   int v2; // ebp
   bot_consolemessage_t *v3;
   char *v4; // eax
